feat: prepared query v1

+++ backend/src/comments/comments.service.ts
@@ -10,22 +10,49 @@ export class CommentsService {
   @DatabaseTable("comment")
   private readonly db: DatabaseService<Comment>;
 
+  securityLevel: number = 3;
+
+  // DONE
   async create(
     createCommentDto: CreateCommentDto,
     user_id: number
   ): Promise<Comment> {
     const now = Date.now();
-    return this.db.insert({
-      query: "content, user_id, event_id, created_at, updated_at",
-      where: `'${createCommentDto.content}', ${user_id}, ${createCommentDto.event_id}, ${now}, ${now}`,
-    });
+    if (this.securityLevel === 1) {
+      return this.db.insert({
+        query: "content, user_id, event_id, created_at, updated_at",
+        where: `'${createCommentDto.content}', ${user_id}, ${createCommentDto.event_id}, ${now}, ${now}`,
+      });
+    } else {
+      const where = `$${1}, $${2}, $${3}, $${4}, $${5}`;
+      return this.db.preparedInsert({
+        query: "content, user_id, event_id, created_at, updated_at",
+        where,
+        variables: [
+          createCommentDto.content,
+          user_id,
+          createCommentDto.event_id,
+          now,
+          now,
+        ],
+      });
+    }
   }
 
+  // DONE
   async findAll(): Promise<Comment[]> {
-    return this.db.queryAll({
-      query: "*",
-      where: "ORDER BY created_at DESC",
-    });
+    if (this.securityLevel === 1) {
+      return this.db.queryAll({
+        query: "*",
+        where: "ORDER BY created_at DESC",
+      });
+    } else {
+      return this.db.preparedQueryAll({
+        query: "*",
+        where: "ORDER BY created_at ASC",
+        variables: [],
+      });
+    }
   }
 
   findOne(id: number) {
@@ -37,9 +64,18 @@ export class CommentsService {
   }
 
   async remove(id: number): Promise<Comment> {
-    return this.db.delete({
-      query: "",
-      where: "id = " + id,
-    });
+    if (this.securityLevel === 1) {
+      return this.db.delete({
+        query: "",
+        where: "id = " + id,
+      });
+    } else {
+      const where = "id = $1";
+      return this.db.preparedDelete({
+        query: "",
+        where,
+        variables: [id],
+      });
+    }
   }
 }
